+++ MyMonitor/JBD.MonitorCozinha.WebAdmin/Models/NumeroPedidoViewModel.cs
@@ -5,7 +5,7 @@ namespace JBD.MonitorCozinha.WebAdmin.Models
 {
     public class NumeroPedidoViewModel
     {
-        public int IdNumeroPedido { get; set; }
+        public Int64 IdNumeroPedido { get; set; }
         public int IdUnidade { get; set; }
         public int IdEmpresa { get; set; }
         public string NumeroPedido { get; set; }
